+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/MetricsAdvisorClient.cs
@@ -671,10 +671,11 @@ namespace Azure.AI.MetricsAdvisor
         /// <param name="feedback">The <see cref="MetricFeedback"/> to be created.</param>
         /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
         /// <returns>
-        /// A <see cref="Response{T}"/> containing the created <see cref="MetricFeedback"/>s.
+        /// A <see cref="Response{T}"/> containing the result of the operation. The result is a <c>string</c>
+        /// containing the ID of the newly created feedback.
         /// </returns>
         /// <exception cref="ArgumentNullException"><paramref name="feedback"/> is null.</exception>
-        public virtual async Task<Response<MetricFeedback>> CreateMetricFeedbackAsync(MetricFeedback feedback, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<string>> CreateMetricFeedbackAsync(MetricFeedback feedback, CancellationToken cancellationToken = default)
         {
             Argument.AssertNotNull(feedback, nameof(feedback));
 
@@ -684,10 +685,9 @@ namespace Azure.AI.MetricsAdvisor
             try
             {
                 ResponseWithHeaders<AzureCognitiveServiceMetricsAdvisorRestAPIOpenAPIV2CreateMetricFeedbackHeaders> response = await _serviceRestClient.CreateMetricFeedbackAsync(feedback, cancellationToken).ConfigureAwait(false);
+                string feedbackId = ClientCommon.GetFeedbackId(response.Headers.Location);
 
-                feedback.Id = ClientCommon.GetFeedbackId(response.Headers.Location);
-
-                return Response.FromValue(feedback, response.GetRawResponse());
+                return Response.FromValue(feedbackId, response.GetRawResponse());
             }
             catch (Exception e)
             {
@@ -702,10 +702,11 @@ namespace Azure.AI.MetricsAdvisor
         /// <param name="feedback">The <see cref="MetricFeedback"/> to be created.</param>
         /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
         /// <returns>
-        /// A <see cref="Response{T}"/> containing the created <see cref="MetricFeedback"/>s.
+        /// A <see cref="Response{T}"/> containing the result of the operation. The result is a <c>string</c>
+        /// containing the ID of the newly created feedback.
         /// </returns>
         /// <exception cref="ArgumentNullException"><paramref name="feedback"/> is null.</exception>
-        public virtual Response<MetricFeedback> CreateMetricFeedback(MetricFeedback feedback, CancellationToken cancellationToken = default)
+        public virtual Response<string> CreateMetricFeedback(MetricFeedback feedback, CancellationToken cancellationToken = default)
         {
             Argument.AssertNotNull(feedback, nameof(feedback));
 
@@ -715,10 +716,9 @@ namespace Azure.AI.MetricsAdvisor
             try
             {
                 ResponseWithHeaders<AzureCognitiveServiceMetricsAdvisorRestAPIOpenAPIV2CreateMetricFeedbackHeaders> response = _serviceRestClient.CreateMetricFeedback(feedback, cancellationToken);
+                string feedbackId = ClientCommon.GetFeedbackId(response.Headers.Location);
 
-                feedback.Id = ClientCommon.GetFeedbackId(response.Headers.Location);
-
-                return Response.FromValue(feedback, response.GetRawResponse());
+                return Response.FromValue(feedbackId, response.GetRawResponse());
             }
             catch (Exception e)
             {