+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/MetricsAdvisorAdministrationClientLiveTests.cs
@@ -65,14 +65,14 @@ namespace Azure.AI.MetricsAdvisor.Tests
                 Description = _dataFeedDescription
             };
 
-            dataFeed = await adminClient.CreateDataFeedAsync(dataFeed).ConfigureAwait(false);
+            string dataFeedId = await adminClient.CreateDataFeedAsync(dataFeed).ConfigureAwait(false);
 
-            Assert.That(dataFeed.Id, Is.Not.Null);
+            Assert.That(dataFeedId, Is.Not.Null);
 
             dataFeed.Description = Recording.GenerateAlphaNumericId("desc");
-            await adminClient.UpdateDataFeedAsync(dataFeed.Id, dataFeed).ConfigureAwait(false);
+            await adminClient.UpdateDataFeedAsync(dataFeedId, dataFeed).ConfigureAwait(false);
 
-            await adminClient.DeleteDataFeedAsync(dataFeed.Id);
+            await adminClient.DeleteDataFeedAsync(dataFeedId);
         }
 
         [RecordedTest]
@@ -86,11 +86,11 @@ namespace Azure.AI.MetricsAdvisor.Tests
                 Description = _dataFeedDescription
             };
 
-            dataFeed = await adminClient.CreateDataFeedAsync(dataFeed).ConfigureAwait(false);
+            string dataFeedId = await adminClient.CreateDataFeedAsync(dataFeed).ConfigureAwait(false);
 
-            Assert.That(dataFeed.Id, Is.Not.Null);
+            Assert.That(dataFeedId, Is.Not.Null);
 
-            DataFeed getDataFeed = await adminClient.GetDataFeedAsync(dataFeed.Id);
+            DataFeed getDataFeed = await adminClient.GetDataFeedAsync(dataFeedId);
 
             getDataFeed.Description = Recording.GenerateAlphaNumericId("desc");
             getDataFeed.MissingDataPointFillSettings.CustomFillValue = 42;
@@ -98,7 +98,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
 
             await adminClient.UpdateDataFeedAsync(getDataFeed.Id, getDataFeed).ConfigureAwait(false);
 
-            await adminClient.DeleteDataFeedAsync(dataFeed.Id);
+            await adminClient.DeleteDataFeedAsync(dataFeedId);
         }
 
         [RecordedTest]
@@ -109,15 +109,15 @@ namespace Azure.AI.MetricsAdvisor.Tests
 
             DataFeed dataFeed = new DataFeed(_blobFeedName, _blobSource, _dailyGranularity, _dataFeedSchema, _dataFeedIngestionSettings);
 
-            dataFeed = await adminClient.CreateDataFeedAsync(dataFeed).ConfigureAwait(false);
+            string dataFeedId = await adminClient.CreateDataFeedAsync(dataFeed).ConfigureAwait(false);
 
-            Assert.That(dataFeed.Id, Is.Not.Null);
+            Assert.That(dataFeedId, Is.Not.Null);
 
             dataFeed.Description = Recording.GenerateAlphaNumericId("desc");
 
-            await adminClient.UpdateDataFeedAsync(dataFeed.Id, dataFeed).ConfigureAwait(false);
+            await adminClient.UpdateDataFeedAsync(dataFeedId, dataFeed).ConfigureAwait(false);
 
-            await adminClient.DeleteDataFeedAsync(dataFeed.Id).ConfigureAwait(false);
+            await adminClient.DeleteDataFeedAsync(dataFeedId).ConfigureAwait(false);
         }
 
         [RecordedTest]
@@ -125,16 +125,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         {
             var adminClient = GetMetricsAdvisorAdministrationClient();
 
-            var config = await CreateMetricAnomalyDetectionConfiguration(adminClient).ConfigureAwait(false);
+            string createdConfigurationId = await CreateMetricAnomalyDetectionConfiguration(adminClient).ConfigureAwait(false);
 
-            AnomalyDetectionConfiguration createdConfiguration = await adminClient.CreateMetricAnomalyDetectionConfigurationAsync(config).ConfigureAwait(false);
+            Assert.That(createdConfigurationId, Is.Not.Null);
 
-            Assert.That(createdConfiguration.Id, Is.Not.Null);
-            Assert.That(createdConfiguration.Name, Is.EqualTo(config.Name));
+            AnomalyDetectionConfiguration getConfig = await adminClient.GetMetricAnomalyDetectionConfigurationAsync(createdConfigurationId).ConfigureAwait(false);
 
-            AnomalyDetectionConfiguration getConfig = await adminClient.GetMetricAnomalyDetectionConfigurationAsync(createdConfiguration.Id).ConfigureAwait(false);
-
-            Assert.That(getConfig.Id, Is.EqualTo(createdConfiguration.Id));
+            Assert.That(getConfig.Id, Is.EqualTo(createdConfigurationId));
 
             getConfig.Description = "updated";
 
@@ -150,7 +147,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
 
             Assert.That(getConfig.Description, Is.EqualTo(userCreatedModel.Description));
 
-            await adminClient.DeleteMetricAnomalyDetectionConfigurationAsync(createdConfiguration.Id).ConfigureAwait(false);
+            await adminClient.DeleteMetricAnomalyDetectionConfigurationAsync(createdConfigurationId).ConfigureAwait(false);
         }
 
         [RecordedTest]
@@ -193,9 +190,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
 
             // Create a Detection Configuration
             DataFeed feed = await GetFirstDataFeed(adminClient);
-            AnomalyDetectionConfiguration detectionConfig = await CreateMetricAnomalyDetectionConfiguration(adminClient).ConfigureAwait(false);
-
-            AnomalyDetectionConfiguration createdAnomalyDetectionConfiguration = await adminClient.CreateMetricAnomalyDetectionConfigurationAsync(detectionConfig).ConfigureAwait(false);
+            string createdAnomalyDetectionConfigurationId = await CreateMetricAnomalyDetectionConfiguration(adminClient).ConfigureAwait(false);
 
             var alertConfigToCreate = new AnomalyAlertConfiguration(
                     Recording.GenerateAlphaNumericId("test"),
@@ -203,7 +198,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
                     new List<MetricAnomalyAlertConfiguration>
                     {
                         new MetricAnomalyAlertConfiguration(
-                            createdAnomalyDetectionConfiguration.Id,
+                            createdAnomalyDetectionConfigurationId,
                             new MetricAnomalyAlertScope(
                                 MetricAnomalyAlertScopeType.TopN,
                                 new DimensionKey(new List<KeyValuePair<string, string>>
@@ -213,22 +208,22 @@ namespace Azure.AI.MetricsAdvisor.Tests
                                 new TopNGroupScope(8, 4, 2)))
                     });
 
-            AnomalyAlertConfiguration createdAlertConfig = await adminClient.CreateAnomalyAlertConfigurationAsync(alertConfigToCreate).ConfigureAwait(false);
+            string createdAlertConfigId = await adminClient.CreateAnomalyAlertConfigurationAsync(alertConfigToCreate).ConfigureAwait(false);
 
-            Assert.That(createdAlertConfig.Id, Is.Not.Null);
+            Assert.That(createdAlertConfigId, Is.Not.Null);
 
             // Validate that we can Get the newly created config
-            AnomalyAlertConfiguration getAlertConfig = await adminClient.GetAnomalyAlertConfigurationAsync(createdAlertConfig.Id).ConfigureAwait(false);
+            AnomalyAlertConfiguration getAlertConfig = await adminClient.GetAnomalyAlertConfigurationAsync(createdAlertConfigId).ConfigureAwait(false);
 
             List<AnomalyAlertConfiguration> getAlertConfigs = new List<AnomalyAlertConfiguration>();
 
-            await foreach (var config in adminClient.GetAnomalyAlertConfigurationsAsync(createdAnomalyDetectionConfiguration.Id))
+            await foreach (var config in adminClient.GetAnomalyAlertConfigurationsAsync(createdAnomalyDetectionConfigurationId))
             {
                 getAlertConfigs.Add(config);
             }
 
-            Assert.That(getAlertConfig.Id, Is.EqualTo(createdAlertConfig.Id));
-            Assert.That(getAlertConfigs.Any(c => c.Id == createdAlertConfig.Id));
+            Assert.That(getAlertConfig.Id, Is.EqualTo(createdAlertConfigId));
+            Assert.That(getAlertConfigs.Any(c => c.Id == createdAlertConfigId));
 
             getAlertConfig.Description = "Updated";
             getAlertConfig.CrossMetricsOperator = MetricAnomalyAlertConfigurationsOperator.And;
@@ -236,7 +231,7 @@ namespace Azure.AI.MetricsAdvisor.Tests
             await adminClient.UpdateAnomalyAlertConfigurationAsync(getAlertConfig.Id, getAlertConfig).ConfigureAwait(false);
 
             // Validate that the update succeeded.
-            getAlertConfig = await adminClient.GetAnomalyAlertConfigurationAsync(createdAlertConfig.Id).ConfigureAwait(false);
+            getAlertConfig = await adminClient.GetAnomalyAlertConfigurationAsync(createdAlertConfigId).ConfigureAwait(false);
 
             Assert.That(getAlertConfig.Description, Is.EqualTo(getAlertConfig.Description));
 
@@ -246,47 +241,50 @@ namespace Azure.AI.MetricsAdvisor.Tests
             await adminClient.UpdateAnomalyAlertConfigurationAsync(getAlertConfig.Id, alertConfigToCreate).ConfigureAwait(false);
 
             // Validate that the update succeeded.
-            getAlertConfig = await adminClient.GetAnomalyAlertConfigurationAsync(createdAlertConfig.Id).ConfigureAwait(false);
+            getAlertConfig = await adminClient.GetAnomalyAlertConfigurationAsync(createdAlertConfigId).ConfigureAwait(false);
 
             Assert.That(getAlertConfig.Description, Is.EqualTo(alertConfigToCreate.Description));
 
             // Cleanup
-            await adminClient.DeleteAnomalyAlertConfigurationAsync(createdAlertConfig.Id).ConfigureAwait(false);
+            await adminClient.DeleteAnomalyAlertConfigurationAsync(createdAlertConfigId).ConfigureAwait(false);
         }
 
         [RecordedTest]
         public async Task HookOperations()
         {
             var adminClient = GetMetricsAdvisorAdministrationClient();
+            var emailHook = new EmailNotificationHook(Recording.GenerateAlphaNumericId("test"), new List<string> { "foo@contoso.com" }) { Description = $"{nameof(EmailNotificationHook)} description" };
 
-            NotificationHook createdEmailHook = await adminClient.CreateHookAsync(new EmailNotificationHook(Recording.GenerateAlphaNumericId("test"), new List<string> { "foo@contoso.com" }) { Description = $"{nameof(EmailNotificationHook)} description" }).ConfigureAwait(false);
+            string createdEmailHookId = await adminClient.CreateHookAsync(emailHook).ConfigureAwait(false);
 
-            EmailNotificationHook getEmailHook = (await adminClient.GetHookAsync(createdEmailHook.Id).ConfigureAwait(false)).Value as EmailNotificationHook;
+            EmailNotificationHook getEmailHook = (await adminClient.GetHookAsync(createdEmailHookId).ConfigureAwait(false)).Value as EmailNotificationHook;
 
             getEmailHook.Description = "updated description";
             getEmailHook.EmailsToAlert.Add($"{Recording.GenerateAlphaNumericId("user")}@contoso.com");
 
             await adminClient.UpdateHookAsync(getEmailHook.Id, getEmailHook).ConfigureAwait(false);
 
-            NotificationHook createdWebHook = await adminClient.CreateHookAsync(new WebNotificationHook(Recording.GenerateAlphaNumericId("test"), "http://contoso.com") { Description = $"{nameof(WebNotificationHook)} description" }).ConfigureAwait(false);
+            var webHook = new WebNotificationHook(Recording.GenerateAlphaNumericId("test"), "http://contoso.com") { Description = $"{nameof(WebNotificationHook)} description" };
 
-            createdWebHook.Description = "updated description";
+            string createdWebHookId = await adminClient.CreateHookAsync(webHook).ConfigureAwait(false);
 
-            await adminClient.UpdateHookAsync(createdEmailHook.Id, createdEmailHook).ConfigureAwait(false);
+            webHook.Description = "updated description";
 
-            WebNotificationHook getWebHook = (await adminClient.GetHookAsync(createdWebHook.Id).ConfigureAwait(false)).Value as WebNotificationHook;
+            await adminClient.UpdateHookAsync(createdEmailHookId, emailHook).ConfigureAwait(false);
+
+            WebNotificationHook getWebHook = (await adminClient.GetHookAsync(createdWebHookId).ConfigureAwait(false)).Value as WebNotificationHook;
 
             getWebHook.Description = "updated description";
             getWebHook.CertificateKey = Recording.GenerateAlphaNumericId("key");
 
             List<NotificationHook> hooks = await adminClient.GetHooksAsync(new GetHooksOptions { HookNameFilter = getWebHook.Name }).ToEnumerableAsync().ConfigureAwait(false);
 
-            Assert.That(getEmailHook.Id, Is.EqualTo(createdEmailHook.Id));
-            Assert.That(getEmailHook.Name, Is.EqualTo(createdEmailHook.Name));
+            Assert.That(getEmailHook.Id, Is.EqualTo(createdEmailHookId));
+            Assert.That(getEmailHook.Name, Is.EqualTo(emailHook.Name));
             Assert.That(hooks, Is.Not.Empty);
-            Assert.That(hooks.Any(h => h.Name == getWebHook.Name), $"hooks should contain name {createdEmailHook.Name}, but contained names: {string.Join(",", hooks.Select(h => h.Name))}");
+            Assert.That(hooks.Any(h => h.Name == getWebHook.Name), $"hooks should contain name {emailHook.Name}, but contained names: {string.Join(",", hooks.Select(h => h.Name))}");
 
-            await adminClient.DeleteHookAsync(createdEmailHook.Id).ConfigureAwait(false);
+            await adminClient.DeleteHookAsync(createdEmailHookId).ConfigureAwait(false);
         }
 
         [RecordedTest]